Early changes
Initial componentizing work for the login dialog
